fix(KEN-631): the plain item's own slot is replaceable, not empty
The occupancy read answered "free" the moment the slot was itself the
plain item, without looking at what else the directory holds. A plain
skill and a namespaced one under the same name coexist by design, so a
slot holding both read as free: the capture trashed the directory,
taking the namespaced item with it and leaving its declaration behind.

Being the plain item now excludes only that item's own files from the
occupancy count. A child the same judge calls an item is still an
occupant, and the refusal names it.

The plain-slot controls move to their own module because the file was
at the test-class ceiling. New fixtures under a tests directory bind
the canonical root, so the moved tests take it; every assertion is the
one it was written with.
